Fixed small issues in ConsAggShock, etc
- Fixed small accounting errors in ConsAggShock, regarding timing of
depreciation and shocks

- Fixed some docstring errors in ConsAggShock and ConsIndShock

- Removed duplicative gridpoint at bottom in ConsAggShock (was added
when I added arbitrary borrowing constraint)

- Fixed integer issue in HARKsimulation.drawDiscrete (issue #116)

+++ HARKsimulation.py
@@ -272,7 +272,7 @@ def drawDiscrete(N,P=[1.0],X=[0.0],exact_match=False,seed=0):
     
     if exact_match:
         events = np.arange(P.size) # just a list of integers
-        cutoffs = np.round(np.cumsum(P)*N) # cutoff points between discrete outcomes
+        cutoffs = np.round(np.cumsum(P)*N).astype(int) # cutoff points between discrete outcomes
         top = 0
         # Make a list of event indices that closely matches the discrete distribution
         event_list        = []
